refactor: remove unused types
