+++ packages/web3-providers/src/types/index.ts
@@ -1,3 +1,4 @@
+// Providers
 export * from './Authorization.js'
 export * from './Bundler.js'
 export * from './DSearch.js'
@@ -26,3 +27,7 @@ export * from './TokenList.js'
 export * from './Trending.js'
 export * from './Twitter.js'
 export * from './Web3.js'
+
+// Implementations
+export * from '../DeBank/type.js'
+export * from '../Zerion/type.js'
