+++ packages/web3-shared/evm/src/types/index.ts
@@ -81,22 +81,6 @@ export enum ChainId {
     Invalid = 0,
 }
 
-/**
- * @deprecated
- */
-export interface Wallet {
-    /** User define wallet name. Default address.prefix(6) */
-    name: string
-    /** The address of wallet */
-    address: string
-    /** yep: removable, nope: unremovable */
-    configurable: boolean
-    /** yep: Mask Wallet, nope: External Wallet */
-    hasStoredKeyInfo: boolean
-    /** yep: Derivable Wallet. nope: UnDerivable Wallet */
-    hasDerivationPath: boolean
-}
-
 export enum AddressType {
     ExternalOwned = 1,
     Contract = 2,
@@ -123,11 +107,17 @@ export interface PriorEIP1559GasConfig {
 
 export type GasConfig = EIP1559GasConfig | PriorEIP1559GasConfig
 
-export interface GasOptionConfig {
-    maxFeePerGas?: number | string
-    maxPriorityFeePerGas?: number | string
-    gasPrice?: number | string
+export interface GasOption {
+    estimatedSeconds: number
+    // eip1559 only
+    estimatedBaseFee?: string
+    baseFeePerGas?: string
+    // note: for prior 1559 it means gasPrice
+    suggestedMaxFeePerGas: string
+    suggestedMaxPriorityFeePerGas: string
 }
+
+export type GasOptionConfig = Omit<GasConfig, 'gas'>
 
 // Learn more for a full list of supported JSON RPC methods
 // https://eth.wiki/json-rpc/API#json-rpc-methods
@@ -185,14 +175,6 @@ export enum TransactionEventType {
     ERROR = 'error',
 }
 
-export interface GasOptions {
-    rapid: number
-    fast: number
-    standard: number
-    slow: number
-    custom: number
-}
-
 export enum DomainProvider {
     ENS = 'ENS',
     UNS = 'UNS',
@@ -203,67 +185,6 @@ export type UnboxTransactionObject<T> = T extends NonPayableTransactionObject<in
     : T extends PayableTransactionObject<infer S>
     ? S
     : T
-
-export enum FilterTransactionType {
-    ALL = 'all',
-    SEND = 'send',
-    RECEIVE = 'receive',
-    CREATE_LUCKY_DROP = 'create_lucky_drop',
-    FILL_POOL = 'fill_pool',
-}
-
-export enum TransactionType {
-    SEND = 'Send',
-    SWAP = 'swap',
-    RECEIVE = 'Receive',
-    TRANSFER = 'transfer',
-    CREATE_LUCKY_DROP = 'create_lucky_drop',
-    CREATE_RED_PACKET = 'create_red_packet',
-    FILL_POOL = 'fill_pool',
-    CLAIM = 'claim',
-    WITHDRAW = 'withdraw',
-    REFUND = 'refund',
-}
-
-export enum DebankTransactionDirection {
-    SEND = 'send',
-    RECEIVE = 'receive',
-}
-
-export enum ZerionTransactionDirection {
-    IN = 'in',
-    OUT = 'out',
-    SELF = 'self',
-}
-
-export type TransactionDirection = ZerionTransactionDirection | DebankTransactionDirection
-
-export interface TransactionPair {
-    name: string
-    symbol: string
-    address: string
-    direction: TransactionDirection
-    amount: number
-}
-
-export interface TransactionGasFee {
-    eth: number
-    usd: number
-}
-
-export enum TransactionStateType {
-    UNKNOWN = 0,
-    /** Wait for external provider */
-    WAIT_FOR_CONFIRMING = 1,
-    /** Hash is available */
-    HASH = 2,
-    /** Receipt is available */
-    RECEIPT = 3,
-    /** Confirmed or Reverted */
-    CONFIRMED = 4,
-    /** Fail to send */
-    FAILED = 5,
-}
 
 /**
  * Keep updating to packages/public-api/src/web.ts
@@ -303,59 +224,6 @@ export enum ProviderType {
     CustomNetwork = 'CustomNetwork',
 }
 
-/**
- * @deprecated use SourceType instead
- */
-export enum FungibleAssetProvider {
-    ZERION = 'Zerion',
-    DEBANK = 'Debank',
-}
-
-/**
- * @deprecated use SourceType instead
- */
-export enum NonFungibleAssetProvider {
-    OPENSEA = 'OpenSea',
-    RARIBLE = 'Rarible',
-    NFTSCAN = 'NFTScan',
-    ZORA = 'Zora',
-}
-
-/**
- * @deprecated
- */
-export type TransactionState =
-    | {
-          type: TransactionStateType.UNKNOWN
-      }
-    | {
-          type: TransactionStateType.WAIT_FOR_CONFIRMING
-
-          /**
-           * @deprecated Don't depend on this property will be removed in the future
-           */
-          hash?: string
-      }
-    | {
-          type: TransactionStateType.HASH
-          hash: string
-      }
-    | {
-          type: TransactionStateType.RECEIPT
-          receipt: TransactionReceipt
-      }
-    | {
-          type: TransactionStateType.CONFIRMED
-          no: number
-          receipt: TransactionReceipt
-          reason?: string
-      }
-    | {
-          type: TransactionStateType.FAILED
-          error: Error & { code?: number }
-          receipt?: TransactionReceipt
-      }
-
 export type Web3 = EVM_Web3
 
 export interface Web3Provider {
@@ -378,15 +246,7 @@ export interface Web3Provider {
 
     removeListener(name: string, listener: (event: any) => void): Web3Provider
 }
-export interface GasOption {
-    estimatedSeconds: number
-    // eip1559 only
-    estimatedBaseFee?: string
-    baseFeePerGas?: string
-    // note: for prior 1559 it means gasPrice
-    suggestedMaxFeePerGas: string
-    suggestedMaxPriorityFeePerGas: string
-}
+
 export type Signature = string
 
 export interface Block {
